+++ accumulators/Cargo.toml
@@ -20,9 +20,6 @@ path = '../algebra'
 [dependencies.zei-crypto]
 path = '../crypto'
 
-[dependencies.zei-utils]
-path = '../utils'
-
 [dependencies.storage]
 git = 'https://github.com/FindoraNetwork/storage.git'
 tag = 'v0.1.4'
