+++ algebra/src/jubjub.rs
@@ -1,23 +1,15 @@
 use crate::{
     bls12_381::BLSScalar,
+    cmp::Ordering,
     errors::AlgebraError,
-    traits::{Group, Scalar},
+    hash::{Hash, Hasher},
+    prelude::*,
 };
 use ark_ec::{AffineCurve, ProjectiveCurve};
 use ark_ed_on_bls12_381::{EdwardsAffine as AffinePoint, EdwardsProjective, Fr};
 use ark_ff::{BigInteger, Field, PrimeField};
 use ark_serialize::{CanonicalDeserialize, CanonicalSerialize};
-use ark_std::{
-    cmp::Ordering,
-    hash::{Hash, Hasher},
-    ops::{Add, AddAssign, Mul, MulAssign, Sub, SubAssign},
-    rand::{CryptoRng, RngCore},
-    One, UniformRand, Zero,
-};
 use digest::{generic_array::typenum::U64, Digest};
-use ruc::*;
-use std::ops::Neg;
-use zei_utils::{derive_prng_from_hash, u8_le_slice_to_u64};
 
 /// The number of bytes for a scalar value over Jubjub
 pub const JUBJUB_SCALAR_LEN: usize = 32;
@@ -367,12 +359,9 @@ impl JubjubPoint {
 mod jubjub_groups_test {
     use crate::{
         jubjub::{JubjubPoint, JubjubScalar},
-        traits::{
-            group_tests::{test_scalar_operations, test_scalar_serialization},
-            Group, Scalar,
-        },
+        prelude::*,
+        traits::group_tests::{test_scalar_operations, test_scalar_serialization},
     };
-    use ark_std::{ops::*, rand::SeedableRng};
     use rand_chacha::ChaCha20Rng;
 
     #[test]
