+++ algebra/src/ristretto.rs
@@ -1,9 +1,4 @@
-use crate::{errors::AlgebraError, traits::Group, traits::Scalar};
-use ark_std::{
-    ops::*,
-    rand::{CryptoRng, RngCore},
-    One, Zero,
-};
+use crate::{errors::AlgebraError, prelude::*};
 use byteorder::ByteOrder;
 use curve25519_dalek::{
     constants::{ED25519_BASEPOINT_POINT, RISTRETTO_BASEPOINT_POINT},
@@ -12,7 +7,6 @@ use curve25519_dalek::{
     traits::Identity,
 };
 use digest::{generic_array::typenum::U64, Digest};
-use ruc::*;
 
 /// The number of bytes for a scalar value over BLS12-381
 pub const RISTRETTO_SCALAR_LEN: usize = 32;
