+++ api/src/anon_xfr/keys.rs
@@ -2,11 +2,9 @@ use rand_core::{CryptoRng, RngCore};
 use ruc::*;
 use wasm_bindgen::prelude::*;
 use zei_algebra::jubjub::{JubjubPoint, JubjubScalar, JUBJUB_SCALAR_LEN};
-use zei_algebra::traits::Group;
+use zei_algebra::prelude::*;
 use zei_crypto::basics::signatures::schnorr;
 use zei_crypto::basics::signatures::schnorr::{KeyPair, PublicKey};
-use zei_utils::errors::ZeiError;
-use zei_utils::serialization::ZeiFromToBytes;
 
 const AXFR_SECRET_KEY_LENGTH: usize = JUBJUB_SCALAR_LEN;
 const AXFR_PUBLIC_KEY_LENGTH: usize = JubjubPoint::COMPRESSED_LEN;
@@ -112,7 +110,7 @@ mod test {
     use crate::anon_xfr::keys::{AXfrKeyPair, AXfrPubKey};
     use rand_chacha::rand_core::SeedableRng;
     use rand_chacha::ChaChaRng;
-    use zei_utils::serialization::ZeiFromToBytes;
+    use zei_algebra::prelude::*;
 
     #[test]
     fn test_axfr_pub_key_serialization() {
