+++ api/src/anon_xfr/structs.rs
@@ -1,19 +1,16 @@
 use crate::anon_xfr::decrypt_memo;
 use crate::anon_xfr::keys::{AXfrKeyPair, AXfrPubKey, AXfrSignature};
 use crate::xfr::structs::{AssetType, OwnerMemo};
-use rand_core::{CryptoRng, RngCore};
-use ruc::*;
 use serde::Serialize;
 use wasm_bindgen::prelude::*;
 use zei_algebra::bls12_381::{BLSPairingEngine, BLSScalar};
 use zei_algebra::jubjub::JubjubScalar;
-use zei_algebra::{traits::Scalar, Zero};
+use zei_algebra::prelude::*;
 use zei_crypto::basics::commitments::rescue;
 use zei_crypto::basics::hybrid_encryption::{
     hybrid_encrypt_with_x25519_key, XPublicKey, XSecretKey,
 };
 use zei_plonk::{plonk::setup::PlonkPf, poly_commit::kzg_poly_com::KZGCommitmentScheme};
-use zei_utils::errors::ZeiError;
 
 pub type Nullifier = BLSScalar;
 pub type Commitment = BLSScalar;
