+++ api/src/api/anon_creds.rs
@@ -1,8 +1,5 @@
-use itertools::Itertools;
-use rand_core::{CryptoRng, RngCore};
-use ruc::*;
 use zei_algebra::bls12_381::{BLSPairingEngine, BLSScalar, BLSG1, BLSG2};
-use zei_algebra::traits::Group;
+use zei_algebra::prelude::*;
 use zei_crypto::anon_creds::{ACCommitOutput, Attribute};
 use zei_crypto::basics::elgamal::elgamal_key_gen;
