+++ api/src/lib.rs
@@ -10,10 +10,7 @@
 #![allow(clippy::many_single_char_names)]
 
 #[macro_use]
-extern crate zei_utils;
-
-#[macro_use]
-extern crate itertools;
+extern crate zei_algebra;
 
 #[macro_use]
 extern crate serde_derive;
@@ -26,5 +23,5 @@ pub mod serialization;
 pub mod setup;
 pub mod xfr;
 
+pub use zei_algebra::errors;
 pub use zei_algebra::ristretto;
-pub use zei_utils::errors;
