+++ api/tests/xfr_note_complex.rs
@@ -1,8 +1,6 @@
 #![deny(warnings)]
 #[cfg(test)]
 pub(crate) mod xfr_note_complex_variable_size {
-    use itertools::Itertools;
-    use rand::SeedableRng;
     use rand_chacha::ChaChaRng;
     use zei::setup::PublicParams;
     use zei::xfr::lib::{gen_xfr_body, verify_xfr_body, XfrNotePoliciesRef};
@@ -10,6 +8,7 @@ pub(crate) mod xfr_note_complex_variable_size {
     use zei::xfr::test_utils::{
         prepare_inputs_and_outputs_with_policies_single_asset, setup_with_policies,
     };
+    use zei_algebra::prelude::*;
 
     fn check_xfr_body(n: usize) {
         let (
