+++ crypto/src/basics/commitments/pedersen.rs
@@ -1,9 +1,9 @@
 use digest::Digest;
 use itertools::Itertools;
 use ruc::*;
+use zei_algebra::errors::ZeiError;
 use zei_algebra::ristretto::RistrettoPoint;
 use zei_algebra::traits::Group;
-use zei_utils::errors::ZeiError;
 
 #[allow(non_snake_case)]
 #[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
