+++ crypto/src/basics/signatures/bls.rs
@@ -1,11 +1,11 @@
 use digest::Digest;
 use rand_core::{CryptoRng, RngCore};
 use ruc::*;
+use zei_algebra::errors::ZeiError;
 use zei_algebra::{
     ops::*,
     traits::{Group, Pairing, Scalar},
 };
-use zei_utils::errors::ZeiError;
 
 type HashFnc = sha2::Sha512;
 
@@ -180,7 +180,7 @@ mod tests {
     use rand_core::SeedableRng;
     use ruc::*;
     use zei_algebra::bls12_381::BLSPairingEngine;
-    use zei_utils::errors::ZeiError;
+    use zei_algebra::errors::ZeiError;
 
     #[test]
     fn bls_signatures() {
