+++ crypto/src/basics/signatures/schnorr.rs
@@ -18,12 +18,12 @@ use rand_core::{CryptoRng, RngCore, SeedableRng};
 use ruc::*;
 use serde::{Deserialize, Serialize};
 use sha2::Sha512;
+use zei_algebra::errors::ZeiError;
+use zei_algebra::serialization::ZeiFromToBytes;
 use zei_algebra::{
     ops::*,
     traits::{Group, Scalar},
 };
-use zei_utils::errors::ZeiError;
-use zei_utils::serialization::ZeiFromToBytes;
 
 #[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
 #[allow(non_snake_case)]
@@ -350,8 +350,8 @@ mod schnorr_sigs {
         use rand_chacha::rand_core::SeedableRng;
         use zei_algebra::jubjub::JubjubPoint;
         use zei_algebra::ristretto::RistrettoPoint;
+        use zei_algebra::serialization::ZeiFromToBytes;
         use zei_algebra::{traits::Group, One};
-        use zei_utils::serialization::ZeiFromToBytes;
 
         fn check_schnorr<G: Group>() {
             let seed = [0_u8; 32];
@@ -426,10 +426,10 @@ mod schnorr_sigs {
             multisig_sign, multisig_verify, KeyPair, MultiSignature, Signature,
         };
         use rand_chacha::rand_core::SeedableRng;
+        use zei_algebra::errors::ZeiError;
         use zei_algebra::jubjub::JubjubPoint;
         use zei_algebra::ristretto::RistrettoPoint;
         use zei_algebra::{traits::Group, One};
-        use zei_utils::errors::ZeiError;
 
         fn check_schnorr_multisig<G: Group>() {
             let seed = [0_u8; 32];
