+++ crypto/src/pedersen_elgamal.rs
@@ -5,15 +5,15 @@ use curve25519_dalek::traits::{Identity, MultiscalarMul};
 use merlin::Transcript;
 use rand_core::{CryptoRng, RngCore};
 use ruc::*;
+use zei_algebra::errors::ZeiError;
 use zei_algebra::ristretto::RistrettoPoint;
 use zei_algebra::ristretto::RistrettoScalar as Scalar;
+use zei_algebra::serialization;
 use zei_algebra::{
     ops::*,
     traits::{Group, Scalar as _},
     One, Zero,
 };
-use zei_utils::errors::ZeiError;
-use zei_utils::serialization;
 
 #[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
 pub struct PedersenElGamalEqProof {
@@ -374,8 +374,8 @@ mod test {
     use ruc::*;
     use serde::de::Deserialize;
     use serde::ser::Serialize;
+    use zei_algebra::errors::ZeiError;
     use zei_algebra::ristretto::{RistrettoPoint, RistrettoScalar};
-    use zei_utils::errors::ZeiError;
 
     #[test]
     fn good_proof_verify() {
