+++ plonk/Cargo.toml
@@ -29,9 +29,6 @@ path = '../algebra'
 [dependencies.zei-crypto]
 path = '../crypto'
 
-[dependencies.zei-utils]
-path = '../utils'
-
 [dependencies.byteorder]
 version = '1.2.3'
 default-features = false
